Move IdentifyAssetDir() to utilities.cpp
