+++ src/Server/shared-server/AppThread.h
@@ -88,41 +88,6 @@ protected:
 	/// \returns true when successful, false otherwise
 	virtual bool LoadConfig(CIni& iniFile);
 
-	/// \brief Represents the source from which an asset directory was resolved.
-	enum class AssetDirSource : uint8_t
-	{
-		/// No valid directory source was found.
-		None,
-
-		/// Directory was provided via the command line.
-		/// This is checked first in the priority order.
-		CommandLine,
-
-		/// Directory was read from the configuration file (i.e. INI file).
-		/// This is checked second in the priority order.
-		Config,
-
-		/// Directory falls back to the hardcoded default directory.
-		/// This will be checked in the current working directory primarily,
-		/// and then fallback to checking the parent directory if not found.
-		/// This is checked last in the priority order.
-		Default
-	};
-
-	/// \brief Identifies an asset directory location, prioritising command-line > INI > defaults.
-	/// \param identifierName		Identifier name of a directory to be used for logging. e.g. "MAP"
-	/// \param commandLineDirectory	The directory passed to the command-line, if applicable. First in priority order.
-	/// \param configDirectory		The directory as stored in the INI config, if applicable. Second in priority order.
-	/// \param defaultDirectory		The default directory location, relative to the current working directory (e.g. ./MAP/).
-	///								As a fallback, will try the directory above (e.g. ../MAP/).
-	///								Last in the priority order.
-	/// \param outputPath			Path to output the final identified directory to.
-	/// \returns true when successful, false otherwise
-	AssetDirSource IdentifyAssetDir(const std::string_view identifierName,
-		const std::filesystem::path& commandLineDirectory,
-		const std::filesystem::path& configDirectory, const std::filesystem::path& defaultDirectory,
-		std::filesystem::path* outputPath);
-
 private:
 	/// \brief Wraps the main startup logic so that error handling can be shared by the caller.
 	/// \param iniFile The loaded application ini file.
